Split rendering the examples into separate plugins
This change splits the example rendering plugins to make them easier to
maintain and add new example types. The YAPSY plugin system used by
Nikola can only load one plugin class from a single module, so the
plugins have to go into three separate modules. The render_example_index
and render_example functions in each module have the same signature but
are specialized for that example type. This is because of how YAPSY
imports plugin code by munging the module name, importing from other
modules in the plugins directory is not possible, so these two functions
are nearly repeated in each plugin. The rendering functions are
implemented at the module level to avoid deep indentation.
