--- plugins/build_examples.plugin
@@ -1,8 +1,0 @@
-[Core]
-Name = build_examples
-Module = build_examples
-
-[Documentation]
-Author = Bryan Weber
-Version = 0.1
-Description = Build the Cantera examples into the output
