+++ plugins/render_python_examples.plugin
@@ -1,0 +1,8 @@
+[Core]
+Name = render_python_examples
+Module = render_python_examples
+
+[Documentation]
+Author = Bryan Weber
+Version = 0.1
+Description = Render the Cantera Python examples into the output
